Chore/add http denylist checks (#4985)
add http denylist checks
