+++ packages/components/nodes/tools/RequestsPut/core.ts
@@ -1,6 +1,7 @@
 import { z } from 'zod'
 import fetch from 'node-fetch'
 import { DynamicStructuredTool } from '../OpenAPIToolkit/core'
+import { checkDenyList } from '../../../src/httpSecurity'
 
 export const desc = `Use this when you want to execute a PUT request to update or replace a resource.`
 
@@ -125,6 +126,9 @@ export class RequestsPutTool extends DynamicStructuredTool {
                 ...(params.headers || {}),
                 ...this.headers
             }
+
+            // Check if URL is allowed by security policy
+            await checkDenyList(inputUrl)
 
             const res = await fetch(inputUrl, {
                 method: 'PUT',
